+++ src/ts/01-functions/index.ts
@@ -11,5 +11,6 @@
  */
 
 export * from './escRegExpURL.js';
+export * from './htmlAttributeString.js';
 export * from './mapToObject.js';
 export * from './trailingSlash.js';