+++ src/ts/02-classes/ElementToggle.ts
@@ -8,6 +8,8 @@
  * @license MIT
  */
 
+// import { VariableInspector } from '../../../node_modules/@maddimathon/utility-typescript/dist/classes/VariableInspector.js';
+
 /**
  * Manages toggle containers made both by the Toggle component and elsewhere.
  * 
@@ -15,14 +17,98 @@
  */
 export class ElementToggle {
 
-    static #openEvent: null | Event = null;
-    static #closeEvent: null | Event = null;
+    /**
+     * Changes some properties and attributes on applicable elements since this
+     * is an invalidly configured toggle element.
+     * 
+     * @since 0.1.0-alpha.7
+     */
+    public static abortNew( container?: HTMLElement | null ) {
+
+        if ( container ) {
+            container.setAttribute( 'data-toggle-container', '' );
+        }
+    }
+
+    /**
+     * Runs a standard init in an HTML document with Toggle components to set up
+     * as instances of this class.
+     * 
+     * @since 0.1.0-alpha.7
+     */
+    public static async init( opts?: Partial<ElementToggle.Opts> ) {
+
+        window.addEventListener(
+            'load',
+            () => document.querySelectorAll(
+                '[data-toggle-container]'
+            ).forEach(
+                ( con ) => con.id && ElementToggle.new( con as HTMLElement, opts )
+            ),
+        );
+    }
+
+    /**
+     * Initiates a single instance asynchronously.
+     * 
+     * @since 0.1.0-alpha.7
+     */
+    public static async new(
+        container: HTMLElement | null,
+        opts?: Partial<ElementToggle.Opts>,
+    ) {
+        const containerID = container?.id;
+
+        // returns
+        if ( !container || !containerID ) {
+            ElementToggle.abortNew( container );
+            return null;
+        }
+
+        const allButtons = document.querySelectorAll(
+            `[data-toggle-primary-control=${ containerID }], [data-toggle-control=${ containerID }]`
+        );
+
+        // returns
+        if ( !allButtons.length ) {
+            ElementToggle.abortNew( container );
+            return null;
+        }
+
+        const primaryButton = document.querySelector(
+            `[data-toggle-primary-control=${ containerID }]`
+        ) as HTMLElement ?? allButtons[ 0 ];
+
+        const content: HTMLElement | null = container.querySelector(
+            `[data-toggle-content=${ containerID }]`
+        );
+
+        // returns - invalid setup that won't work
+        if ( !content ) {
+            ElementToggle.abortNew( container );
+            return null;
+        }
+
+        return new ElementToggle(
+            {
+                container,
+                primaryButton,
+                allButtons: Array.from( allButtons ) as [ HTMLElement ] & HTMLElement[],
+                content,
+            },
+            opts,
+        );
+    }
+
+
+    public static openEvent: Event | null = null;
+    public static closeEvent: Event | null = null;
 
 
     /** 
      * @param string  A CSS time value to convert to milliseconds.
      */
-    static cssTimeToMilliseconds( string: string ) {
+    public static cssTimeToMilliseconds( string: string ) {
 
         if ( string.includes( 'ms' ) ) {
             return Number( string.replace( /\s*ms\s*$/gi, '' ) );
@@ -30,72 +116,101 @@ export class ElementToggle {
         return Number( string.replace( /\s*s\s*$/gi, '' ) ) * 1000;
     }
 
-    static createCustomEvents() {
+    public static createCustomEvents() {
 
-        if ( this.#openEvent === null ) {
-            this.#openEvent = new Event( 'toggle-open' );
+        if ( this.openEvent === null ) {
+            this.openEvent = new Event( 'toggle-open' );
         }
 
-        if ( this.#closeEvent === null ) {
-            this.#closeEvent = new Event( 'toggle-close' );
+        if ( this.closeEvent === null ) {
+            this.closeEvent = new Event( 'toggle-close' );
         }
     }
 
 
-    #primaryButton: HTMLElement | null = null;
-    #allButtons: HTMLElement[] | null = null;
-    #container: HTMLElement | null = null;
 
-    /** @type {NodeJS.Timeout|null} */
-    #closingTimeout: NodeJS.Timeout | null = null;
+    /* LOCAL PROPS
+     * ====================================================================== */
 
-    /** in milliseconds @type {number} */
-    #closingTime: number = 1800;
+    /** 
+     * Optional configuration, if any.
+     * @since 0.1.0-alpha.7
+     */
+    protected readonly opts: ElementToggle.Opts;
+
+    /** 
+     * The unique ID for the toggle container to set up.
+     */
+    protected readonly container: HTMLElement;
+
+    protected readonly content: HTMLElement;
+    protected readonly primaryButton: HTMLElement;
+    protected readonly allButtons: HTMLElement[];
+
+    /**
+     * @since 0.1.0-alpha.7
+     */
+    protected readonly containerID: string;
+
+    protected closingTimeout: NodeJS.Timeout | null = null;
+
+    /** 
+     * In milliseconds.
+     */
+    protected closingTime: number;
+
+
+
+    /* CONSTRUCTOR
+     * ====================================================================== */
 
     /** 
      * Class constructor.
-     * 
-     * @param containerID  The unique ID for the toggle container to set up.
      */
-    constructor ( containerID: string ) {
+    public constructor (
 
-        this.#container = document.querySelector( `[data-toggle-container]#${ containerID }` );
+        elements: {
+            container: HTMLElement,
+            primaryButton: HTMLElement,
+            allButtons: [ HTMLElement ] & HTMLElement[],
+            content: HTMLElement,
+        },
+
+        /** Optional configuration, if any. */
+        partialOpts?: Partial<ElementToggle.Opts>,
+    ) {
+        this.opts = {
+            closeWhenUntargetted: false,
+            closingTime: 1800,
+            openWhenTargetted: true,
+            ...partialOpts,
+        };
+
+        this.closingTime = this.opts.closingTime;
+
+        this.container = elements.container;
+        this.primaryButton = elements.primaryButton;
+        this.allButtons = elements.allButtons;
+        this.content = elements.content;
+
+        this.containerID = this.container.id;
 
         // returns
-        if ( !this.#container ) {
-            this.#abortConstructor();
+        if ( !this.container || !this.primaryButton || !this.containerID || !this.content ) {
+            this.abortConstructor();
             return;
         }
 
-        const allButtons = this.#container.querySelectorAll( `[data-toggle-primary-control=${ containerID }], [data-toggle-control=${ containerID }]` );
+        this.toggle = this.toggle.bind( this );
+        this.handleHashChange = this.handleHashChange.bind( this );
 
-        // returns
-        if ( !allButtons.length ) {
-            this.#abortConstructor();
-            return;
-        }
+        const contentID = this.content.id;
 
-        this.#allButtons = Array.from( allButtons ) as HTMLElement[];
-
-        const primaryButton = this.#container.querySelector( `[data-toggle-primary-control=${ containerID }]` );
-
-        this.#primaryButton = primaryButton as HTMLElement ?? this.#allButtons[ 0 ];
-
-        const content = this.#container.querySelector( `[data-toggle-content=${ containerID }]` );
-
-        // returns - invalid setup that won't work
-        if ( !content || !this.#allButtons.length ) {
-            this.#abortConstructor();
-            return;
-        }
-
-        const contentID = content.id;
-
-        content.setAttribute( 'aria-labelledby', this.#primaryButton.id );
-        content.setAttribute( 'role', 'region' );
+        this.content.setAttribute( 'aria-labelledby', this.primaryButton.id );
+        this.content.setAttribute( 'role', 'region' );
 
         if ( contentID ) {
-            this.#allButtons.forEach( ( button ) => {
+            this.allButtons.forEach( ( button ) => {
 
                 if (
                     button.getAttribute( 'role' ) == 'button'
@@ -107,72 +222,164 @@ export class ElementToggle {
             } );
         }
 
-        this.#closingTime = ElementToggle.cssTimeToMilliseconds(
-            getComputedStyle( this.#container ).getPropertyValue( '--toggle-closing-time' )
+        this.closingTime = ElementToggle.cssTimeToMilliseconds(
+            getComputedStyle( this.container ).getPropertyValue( '--toggle-closing-time' )
         );
 
-        const defaultIsOpen = this.#container.hasAttribute( 'data-toggle-container' )
-            && this.#container.getAttribute( 'data-toggle-container' ) === 'open';
+        const isCurrentAnchorTarget = this.opts.openWhenTargetted
+            && this.checkUrlTarget( new URL( window.location.href ) );
 
-        this.toggle = this.toggle.bind( this );
+        const defaultIsOpen =
+            this.container.getAttribute( 'data-toggle-container' ) === 'open'
+            || isCurrentAnchorTarget;
 
-        this.#allButtons.forEach( ( button ) => {
+        this.allButtons.forEach( ( button ) => {
             button.addEventListener( 'click', this.toggle );
 
             if ( button.getAttribute( 'aria-controls' ) ) {
                 button.removeAttribute( 'aria-disabled' );
-                button.setAttribute( 'aria-expanded', 'false' );
+                button.setAttribute( 'aria-expanded', defaultIsOpen ? 'true' : 'false' );
             }
         } );
 
-
         if ( defaultIsOpen ) {
-            this.#open();
+
+            if ( isCurrentAnchorTarget ) {
+                this.openAsTargetAnchor();
+            } else {
+                this.open();
+            }
         } else {
-            this.#container.setAttribute( 'data-toggle-container', 'closed' );
+            this.container.setAttribute( 'data-toggle-container', 'closed' );
+        }
+
+        if ( !isCurrentAnchorTarget ) {
+            this.primaryButton.removeAttribute( 'data-toggle-is-current-target' );
+        }
+
+        if ( this.opts.openWhenTargetted ) {
+            window.addEventListener( 'hashchange', this.handleHashChange );
         }
     }
 
     /**
-     * Changes some properties and attributes on applicable elements since this
-     * is an invalidly configured toggle element.
+     * {@inheritDoc ElementToggle.abortNew}
+     * 
+     * @since 0.1.0-alpha
      */
-    #abortConstructor() {
-
-        if ( this.#container ) {
-            this.#container.setAttribute( 'data-toggle-container', '' );
-        }
+    protected abortConstructor() {
+        ElementToggle.abortNew( this.container );
     }
 
+
+
+    /* UTILITIES
+     * ====================================================================== */
 
     /**
      * Clears the related timeout, if any.
      */
-    #clearTimeout() {
+    protected clearTimeout() {
         /*
          * Clear any timeout currently running (like if someone clicks the
          * button before it's done).
          */
-        if ( this.#closingTimeout !== null ) {
-            clearTimeout( this.#closingTimeout );
+        if ( this.closingTimeout !== null ) {
+            clearTimeout( this.closingTimeout );
         }
     }
 
+    /**
+     * Checks the current url anchor target and checks if it matches the id of
+     * this toggle element.
+     *
+     * @since 0.1.0-alpha.7
+     */
+    protected checkUrlTarget( url: URL ): boolean {
+        // returns
+        if ( !url.hash ) {
+            return false;
+        }
+
+        const hashAsId = url.hash.replace( /^#/gi, '' );
+
+        return hashAsId.toLowerCase() === this.containerID.toLowerCase();
+    }
+
+    /**
+     * If applicable (by opts), checks if the current url anchor targets
+     * this toggle and if so, opens it.
+     *
+     * @since 0.1.0-alpha.7
+     */
+    public handleHashChange( event: HashChangeEvent ) {
+        // returns
+        if ( !this.opts.openWhenTargetted ) {
+            return;
+        }
+
+        const isNewTarget = this.checkUrlTarget( new URL( event.newURL ) );
+
+        if ( !isNewTarget ) {
+            this.primaryButton.removeAttribute( 'data-toggle-is-current-target' );
+        }
+
+        if ( isNewTarget ) {
+            this.openAsTargetAnchor();
+        }
+
+        if (
+            !isNewTarget
+            && this.opts.closeWhenUntargetted
+            && this.checkUrlTarget( new URL( event.oldURL ) )
+            && this.container.getAttribute( 'data-toggle-container' ) === 'open'
+        ) {
+            this.close();
+        }
+    }
+
+    /**
+     * Opens the toggle element as if this is the current url anchor target. 
+     * Opens regardless of the current `this.opts.openWhenTargetted` value.
+     *
+     * @since 0.1.0-alpha.7
+     */
+    protected openAsTargetAnchor() {
+        this.open();
+
+        this.primaryButton.setAttribute( 'data-toggle-is-current-target', 'true' );
+
+        this.primaryButton.addEventListener(
+            'blur',
+            () => this.primaryButton.removeAttribute( 'data-toggle-is-current-target' ),
+            { once: true },
+        );
+
+        this.primaryButton.focus( {
+            // @ts-expect-error
+            focusVisible: true,
+        } );
+    }
+
+
+
+    /* TOGGLING
+     * ====================================================================== */
 
     /**
      * Toggles the open/close state of the element.
      */
-    toggle() {
-        if ( !this.#container ) { return; }
+    public toggle() {
+        if ( !this.container ) { return; }
 
         /*
          * Grab the current state and trigger an opening or closing function!
          */
-        switch ( this.#container.getAttribute( 'data-toggle-container' ) ) {
+        switch ( this.container.getAttribute( 'data-toggle-container' ) ) {
 
             case 'closed':
-                this.#clearTimeout();
-                this.#open();
+                this.clearTimeout();
+                this.open();
                 break;
 
             case 'closing':
@@ -180,8 +387,8 @@ export class ElementToggle {
 
             case 'open':
             default:
-                this.#clearTimeout();
-                this.#close();
+                this.clearTimeout();
+                this.close();
                 break;
         }
     }
@@ -189,54 +396,93 @@ export class ElementToggle {
     /**
      * Toggles the element open.
      */
-    #open() {
-        if ( !this.#allButtons ) { return; }
-        if ( !this.#container ) { return; }
+    protected open() {
+        if ( !this.allButtons ) { return; }
+        if ( !this.container ) { return; }
 
-        this.#closingTime = ElementToggle.cssTimeToMilliseconds(
-            getComputedStyle( this.#container ).getPropertyValue( '--toggle-closing-time' )
+        this.closingTime = ElementToggle.cssTimeToMilliseconds(
+            getComputedStyle( this.container ).getPropertyValue( '--toggle-closing-time' )
         );
 
-        this.#container.setAttribute( 'data-toggle-container', 'open' );
+        this.container.setAttribute( 'data-toggle-container', 'open' );
 
-        this.#allButtons.forEach( ( button ) => {
+        this.allButtons.forEach( ( button ) => {
             if ( button.getAttribute( 'aria-controls' ) ) {
                 button.setAttribute( 'aria-expanded', 'true' );
             }
         } );
 
         ElementToggle.createCustomEvents();
-        this.#container.dispatchEvent( ElementToggle.#openEvent as Event );
+        this.container.dispatchEvent( ElementToggle.openEvent as Event );
     }
 
     /**
      * Toggles the element closed.
      */
-    #close() {
-        if ( !this.#allButtons ) { return; }
-        if ( !this.#container ) { return; }
+    protected close() {
+        if ( !this.allButtons ) { return; }
+        if ( !this.container ) { return; }
 
         /*
          * Adjust the data-toggle-container on the container and the aria-expanded for
          * the button.
          */
-        this.#allButtons.forEach( ( button ) => {
+        this.allButtons.forEach( ( button ) => {
             if ( button.getAttribute( 'aria-controls' ) ) {
                 button.setAttribute( 'aria-expanded', 'false' );
             }
         } );
-        this.#container.setAttribute( 'data-toggle-container', 'closing' );
+        this.container.setAttribute( 'data-toggle-container', 'closing' );
 
         /*
          * Wait for animations to finish.
          */
-        this.#closingTimeout = setTimeout( () => {
+        this.closingTimeout = setTimeout( () => {
             // Sets the data-toggle-container to closed now that animations are over.
-            if ( !this.#container ) { return; }
-            this.#container.setAttribute( 'data-toggle-container', 'closed' );
-        }, this.#closingTime + 50 );
+            if ( !this.container ) { return; }
+            this.container.setAttribute( 'data-toggle-container', 'closed' );
+        }, this.closingTime + 50 );
 
         ElementToggle.createCustomEvents();
-        this.#container.dispatchEvent( ElementToggle.#closeEvent as Event );
+        this.container.dispatchEvent( ElementToggle.closeEvent as Event );
+    }
+}
+
+/**
+ * Utilities for the {@link ElementToggle} class.
+ * 
+ * @since 0.1.0-alpha.7
+ */
+export namespace ElementToggle {
+
+    /**
+     * Options for the configuration of {@link ElementToggle} instances.
+     * 
+     * @since 0.1.0-alpha.7
+     */
+    export interface Opts {
+
+        /**
+         * Whether toggles should close when they are no longer the target of
+         * the url's anchor.
+         *
+         * @default false
+         */
+        closeWhenUntargetted: boolean;
+
+        /**
+         * Default toggle closing time.
+         * 
+         * @default 1800
+         */
+        closingTime: number;
+
+        /**
+         * Whether toggles should open when they are the target of the url's
+         * anchor.
+         * 
+         * @default true
+         */
+        openWhenTargetted: boolean;
     }
 }